requirements.txt
phasepack
wheel
Cython
dotmap
matplotlib
Pillow
scikit-image
scikit-learn
opencv-python-headless
pyfftw
pyyaml
# Extra index for torch and torchvision.
--extra-index-url https://download.pytorch.org/whl/cu116
torch
torchvision
torchmetrics
comet_ml
pytorch-ignite
sporco
scipy

# Must be installed separetely.
# or with `cat requirements.txt | xargs -n 1 -L 1 pip install`
#git+https://github.com/qubvel/segmentation_models.pytorch
#git+https://github.com/waspinator/coco.git@2.1.0
#git+https://github.com/waspinator/pycococreator.git@0.2.0
#gimp-labeling-converter
